a couple minor edits
why did you make it so that if you say a number less than one it tells you "You can't guess a number less than one" but if you say a number greater than the maximum number it can be it just repeats the previous line? It is because they may have forgotten what the range is? (i.e someone deletes what the emu bot said and they don't have the best memory) or is it because you might as well? I dunno, i guess it does make sense though.