Added a backup when editing files
The way it works is:
-It copies the main file
-Writes to the new file
-Deletes the old file
-Renames the new file to the old file
So before I start up the Emu Bot after it gets interrupted by an error or a power outage or something else, I can check if one file (the main file) is present, two files (the main & the new file). If both are present, it didn't have time to finish writing the file and that file is essentially corrupted, so I'll delete that file and continue on knowing that the main file is free of corruption.