Deleted comment and corrected message
Akiva, I don't think you understand how this works.
https://github.com/CapClumsy/emubot/blob/f8b2c19999f07961a9576af9faa325a2c9e021c7/Prototypes/giveprototype.py#L17-L21
Lines 17 and 18 make a uidstr readable and a full integer without an ! at the beginning (a mention in discord is displayed differently in the GUI, but in the code of the message (sort of), a mention is displayed between arrow brackets with an @ symbol and then maybe an ! then the user ID (ex. <@1234567890> or <@!1234567890>)). These lines make it so that it can get rid of an !. You can stop correcting things if you dn't know what they do.