Fixed #20 stats adding issue
It used to set all stats to zero except for the stats you were adding in the for loop that was initially meant to keep the stats only to the three values. But now with the error statement that will make sure that only the three values will be entered.
I'm not sure why it was going having a KeyError though. But that might not be a problem. It might be a thing that JSON does.