Fixed attack cooldown
Yeah I know this is like the 18th time I've fixed the cooldown but I'm getting into stuff I've never done before. I looked at this and wondered how it could possibly work today, and then when I tested it, it didn't! Anyway the problem is that I never set ATTACKED to True if you actually attacked and in the after_invoke it didn't check , it only checked . I think I must've just not finished it.