Added stats backup
Similar to the old theoretical stats backup that was never implemented, but instead of making a backup for every edit it used a task to back it up every 30 seconds. I decided to do this after the odd bug that completely wiped the file clean if an outage happened during file writing, and this should negate that.